Refactor and generalize debug interface a bit
